+++ src/components/ItemList.js
@@ -1,10 +1,10 @@
-import React, { useContext } from 'react'
-import { ItemsContext } from '../context/ItemsContext';
+
+import { getItems } from '../utils/itemsHandler';
 import Item from './Item'
 
 const ItemsList = () => {
-  const { items, dispatch } = useContext(ItemsContext);
-  console.log(items)
+  const items = getItems()
+  items.sort((a, b) => { return new Date(a.createdAt) - new Date (b.createdAt)});
   return (
     <div className='m-16'>
       <ul>
