+++ src/utils/fetcher.js
@@ -10,18 +10,21 @@ export const fetcher = async (prompt)=>{
     }
 
     const url = process.env.REACT_APP_OPENAI_URL;
-    console.log(url);
-    const response = await fetch(url,
-        {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${process.env.REACT_APP_OPENAI_SECRET}`
-            },
-            body: JSON.stringify(payload)
-        });
-    const data = await response.json();
-    console.log(data);
-    return data;
+    try{
+
+        const response = await fetch(url,
+            {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    Authorization: `Bearer ${process.env.REACT_APP_OPENAI_SECRET}`
+                },
+                body: JSON.stringify(payload)
+            });
+            const data = await response.json();
+            return data;
+        }catch (e){
+            console.log(e);
+        }
 
 }
